fix: add validation on load method and restrict deleteItems execution
- Add validation in the  method to check memory data consistency.
- Prevent  method from executing when not using a local database.

+++ library/modules/entities/collection/local-provider/index.ts
@@ -171,6 +171,7 @@ export /*bundle*/ class CollectionLocalProvider extends ReactiveModel<Collection
 
 	// Method to delete multiple items by their IDs using an existing Dexie store
 	async deleteItems(ids) {
+		if (!this.#apply || !this.#localdb) return;
 		try {
 			await this.store.bulkDelete(ids);
 		} catch (error) {
